+++ DearPyGui/src/core/AppItems/mvAppItem.h
@@ -293,6 +293,7 @@ namespace Marvel {
         // other than drawing.
         //-----------------------------------------------------------------------------
         virtual void customAction() {};
+        virtual void alternativeCustomAction() {};
 
         //-----------------------------------------------------------------------------
         // These methods handle setting the widget's value using PyObject*'s or
@@ -369,6 +370,7 @@ namespace Marvel {
         const std::string&                  getSpecifiedLabel() const { return m_specificedlabel; }
         mvAppItem*                          getRoot() const;
         int                                 getLocation() const { return m_location; }
+        bool                                isAltCustomActionRequested() const { return m_triggerAlternativeAction; }
 
     protected:
 
@@ -407,6 +409,7 @@ namespace Marvel {
         ImVec2         m_previousCursorPos = { 0.0f, 0.0f };
         int            m_location = -1;
         std::string    m_label; // internal label
+        bool           m_triggerAlternativeAction = false;
 
         mvAppItem*                    m_parentPtr = nullptr;