+++ DearPyGui/src/core/AppItems/themes/mvTheme.h
@@ -32,8 +32,13 @@ namespace Marvel {
 
 		mvTheme(mvUUID uuid);
 
+		// overriding until we can remove these
+		bool preDraw() override { return true; }
+		void postDraw() override {}
+
 		void draw(ImDrawList* drawlist, float x, float y) override;
 		void customAction() override;
+		void alternativeCustomAction() override;
 		bool canChildBeAdded(mvAppItemType type) override;
 		void handleSpecificKeywordArgs(PyObject* dict) override;